+++ src/components/CheckoutForm.js
@@ -1,5 +1,7 @@
 import React, { useState } from "react";
 
+import useForm from "../hooks/useForm";
+
 const initialValue = {
   firstName: "",
   lastName: "",
@@ -14,17 +16,7 @@ const initialValue = {
 // and replace the necessary stateful logic from CheckoutForm with the hook
 
 const CheckoutForm = (props) => {
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-  const [values, setValues] = useState(initialValue);
-
-  const handleChanges = (e) => {
-    setValues({ ...values, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowSuccessMessage(true);
-  };
+  const [values, showSuccessMessage, handleChanges, handleSubmit] = useForm(initialValue);
 
   return (
     <>
